+++ test/cases/js-import-css-dynamic-import/expected/120.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[120],{120:(e,s,c)=>{c.r(s),console.log(">> component a")}}]);