+++ test/cases/js-import-css-dynamic-import/expected/851.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[851],{851:(e,n,c)=>{c.e(942).then(c.t.bind(c,942,23)),console.log(">> component c")}}]);