+++ test/cases/js-import-css-dynamic-import/expected/942.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[942],{942:()=>{}}]);