+++ test/cases/js-import-css-dynamic-import/src/component-a.js
@@ -1,0 +1,4 @@
+import './component-a1.css';
+import './component-a2.css';
+
+console.log('>> component a');