+++ test/cases/js-import-css-dynamic-import/src/component-b.js
@@ -1,0 +1,8 @@
+// dynamic load styles
+const loadStyles = () => import('./component-b1.css');
+loadStyles();
+
+// normal import
+//import './component-b1.css';
+
+console.log('>> component b');