+++ test/cases/js-import-css-dynamic-import/src/component-c.js
@@ -1,0 +1,4 @@
+// dynamic import of the style in the dynamic imported js
+import('./component-c1.css');
+
+console.log('>> component c');